requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-embed-video==1.4.10
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2
django-heroku==0.3.1
pyscopg2==2.9.1